requirements.txt
django
djangorestframework
Pillow
numpy
scikit-learn
tensorflow
djangorestframework
drf-yasg